get_pb: BREAKING -> magisk
- disable debug force gpg download
- new approach using completely untouched magisk patch scripts

This requires 3 changes, all in your mkbootimg.mk:

1. `ROOT_BOOT_BIN` must be changed from `$(OUT_DIR)/.magisk/root_boot.sh` to `$(OUT_DIR)/.magisk/boot_patch.sh`
2. `MAGISK_TARGET_ARCH=arm64 $(ROOT_BOOT_BIN) $$PWD/$@` becomes `@/bin/bash $(ROOT_BOOT_BIN) $$PWD/$@`
3. `@cp -v $(OUT_DIR)/.magisk/IMAGES/boot.img  $(PRODUCT_OUT)/boot.img` becomes `@cp -v $(OUT_DIR)/.magisk/new-boot.img  $(PRODUCT_OUT)/boot.img`

and optionally when using 32bit devices (ONLY then):

1. change your `device/<vendor>/<model>/vendorsetup.sh` and add: `MAGISK_TARGET_ARCH=arm`

Example:
https://github.com/sfX-android/android_device_oneplus_hotdog/commit/4adaab24fccaf94773c3394365d2465cb7c68932

Also new is that `MAGISK_TARGET_ARCH` will be parsed when set in your `vendorsetup.sh` or specified as env variable.
It can be `arm` or `arm64` which will copy the correct libs according to whats specified.
If unset the default is expecting `arm64` - so ensure you set it for 32bit devices (see above)!

Signed-off-by: steadfasterX <steadfasterX@gmail.com>