mk: support strict artifact_path_requirements:
if strict checking is defined for a device the build fails bc of artifact_path_requirements.mk.
this is bc extendrom might write (depends on the users setup) into:

- system/app
- system/priv-app
- system/etc/init
- system/etc/default-permissions
- system/etc/permissions
- system/etc/sysconfig

these paths are (actually whole system/ is) protected on newer Android versions and instead system_ext should be used for system near stuff - which is what we do now.

fixes:

> build/make/core/artifact_path_requirements.mk:30: warning:  device/google/bluejay/lineage_bluejay.mk produces files inside build/make/target/product/generic_system.mks artifact path requirement.
> Offending entries:
> [...]
> build/make/core/artifact_path_requirements.mk:30: error: Build failed.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>